pyproject.toml
[tool.poetry]
package-mode = false
name = "hands-be"
version = "0.1.0"
description = ""
authors = ["seungwon0113 <jungseungwon0113@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-restframework = "^0.0.1"
drf-spectacular = "^0.28.0"
environ = "^1.0"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-stubs = "^3.15.2"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"
django-cors-headers = "^4.6.0"
gunicorn = "^23.0.0"
pillow = "^11.0.0"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.159.0"

[tool.django-stubs]
django_settings_module = "config.settings.base"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pre-commit = "^4.0.1"
coverage = "^7.6.9"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py313"]
include = ".pyi?$"
extend-exclude = '''
/(
    .eggs
  | .git
  | .mypy_cache
  | .venv
  | venv
  | env
  | _build
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".gitignore", ".dockerignore"]
skip_glob = ["*/migrations/*"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
